Move TxOutToPublicCoin from zpivchain.cpp to zpivmodule.cpp
it's only used there and has nothing to do with the "zpivchain" (whatever that was).

Plus clean unneeded dependencies.
