+++ src/zpivchain.h
@@ -5,24 +5,12 @@
 #ifndef PIVX_ZPIVCHAIN_H
 #define PIVX_ZPIVCHAIN_H
 
-#include "chain.h"
-#include "libzerocoin/Coin.h"
-#include "libzerocoin/Denominations.h"
 #include "libzerocoin/CoinSpend.h"
-#include <list>
-#include <string>
 
-class CBlock;
-class CBlockIndex;
 class CBigNum;
-class CTransaction;
 class CTxIn;
-class CTxOut;
-class CValidationState;
-class uint256;
 
 bool IsSerialInBlockchain(const CBigNum& bnSerial, int& nHeightTx);
 libzerocoin::CoinSpend TxInToZerocoinSpend(const CTxIn& txin);
-bool TxOutToPublicCoin(const CTxOut& txout, libzerocoin::PublicCoin& pubCoin, CValidationState& state);
 
 #endif //PIVX_ZPIVCHAIN_H
